fixed the infinity loop error but got one new one
error was caused by the constructor constantly calling the forecast function
new error :
have to use a extra function to pare the data into an array, I could use a global thing or make it every single file. I think I am going to go with a global thing so that it only get the data once and this on load.